BT29 Crosshair Firingstates!
Now the crosshair fully changes color as you are in different Firing states. Those firing states are: Reloading, Aiming, Locked.
When your gun is reloading, it'll show up as red, if the turret or barrel is still turning to a new desired position or the projectile cannot land on set location, it'll show as Aiming and if the area is hittable and the turret nor' barrel are moving, it'll show green as Locked.