BT49 Repaired a major issue.
Unreal did some magic which for some reason completely broke the tank blueprint, so I had to remake it from scratch. With all the components already made it was more of a lego assembling then anything else, which isn't that bad I suppose. I also fixed the AI navmesh, so that all works now.
Next up is really just to balance out the values and make sure the tank cannot accelerate when not touching the ground.